ml-service/requirements.txt
# DrishtiFlow Prediction Engine - Python Dependencies
# ===================================================

# Web Framework
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.18

# ML Core
xgboost==2.1.3
scikit-learn==1.5.2
numpy==2.2.1
pandas==2.2.3

# ONNX Runtime for fast inference
onnxruntime==1.20.1
onnxmltools==1.12.0
skl2onnx==1.17.0

# Feature Engineering & Data
scipy==1.15.0

# Caching
cachetools==5.5.1

# Logging & Monitoring
loguru==0.7.3

# Environment
python-dotenv==1.0.1

# CORS
pydantic==2.10.4

# HTTP client (for external signals)
httpx==0.28.1

# Testing
pytest==8.3.4
